requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.26

# LLM Provider (Google Gemini)
langchain-google-genai>=1.0.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional (for development and testing)
pytest>=7.4.0
